split implementation of INumber<TSelf> and INumberBase<TSelf> to individual files
