+++ src/Smdn.Fundamental.UInt24n/Smdn/UInt24.cs
@@ -64,10 +64,14 @@ public partial struct UInt24 {
 #pragma warning restore IDE0060
 
 #if FEATURE_GENERIC_MATH
-  private static bool TryCreateCore<TOther>(
+  /*
+   * INumberBase<TSelf>
+   */
+  private static bool TryConvertFromTruncating<TOther>(
     TOther value,
+    bool throwIfOverflow,
     out UInt24 result
-  ) where TOther : INumber<TOther>
+  ) where TOther : INumberBase<TOther>
   {
     result = default;
 
@@ -106,10 +110,10 @@ public partial struct UInt24 {
       val = unchecked((int)((decimal)(object)value));
 #pragma warning restore IDE0045
     else
-      throw UInt24n.CreateTypeIsNotConvertibleException<UInt24, TOther>();
+      return false; // is not convertible
 
-    if (val is < minValue or > maxValue)
-      return false; // overflow
+    if (throwIfOverflow && val is < minValue or > maxValue)
+      throw UInt24n.CreateOverflowException<UInt24>(value);
 
     result = new(unchecked((uint)val));
